requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.9.1+cpu
torch==1.8.1+cpu 
gunicorn==20.0.4
Werkzeug==0.16.0
Flask==1.1.2
requests==2.24.0
Pillow==8.4.0
